Move NodeId out of api
